+++ client/src/js/cookieconsent.js
@@ -10,7 +10,6 @@ import Cookies from "js-cookie";
             const cookieExpiry = dataElement.getAttribute('data-expiry');
             const additionalHostLinks = dataElement.getAttribute('data-additional-host-links');
 
-
             const popup = document.getElementById('CookieConsent');
             if (typeof (popup) != 'undefined' && popup != null) {
                 if (document.cookie.match(new RegExp('(^| )' + cookieName + '=([^;]+)'))) {
@@ -23,10 +22,7 @@ import Cookies from "js-cookie";
                 Array.prototype.forEach.call(buttons, function (button) {
                     button.addEventListener('click', function (e) {
                         e.preventDefault();
-                        const xhr = new XMLHttpRequest();
-                        xhr.open('GET', this.href);
-                        xhr.setRequestHeader('x-requested-with', 'XMLHttpRequest');
-                        xhr.send();
+                        // get cookie groups to be set
                         const newCookieGroups = this.getAttribute('data-cookie-groups');
                         // check for existing settings and merge them
                         let updatedCookieGroups = newCookieGroups;
@@ -37,8 +33,23 @@ import Cookies from "js-cookie";
                             const updatedCookieGroupsArray = [...new Set([...currentCookieGroupsArray, ...newCookieGroupsArray])];
                             updatedCookieGroups = updatedCookieGroupsArray.join(",");
                         }
-                        // set cookie
-                        Cookies.set(cookieName, updatedCookieGroups, { path: '/', expires: parseInt(cookieExpiry)})
+                        // send cookie setting request to backend
+                        const xhr = new XMLHttpRequest();
+                        xhr.open('GET', this.href);
+                        xhr.setRequestHeader('x-requested-with', 'XMLHttpRequest');
+                        xhr.onload = function() {
+                            if (xhr.status >= 200 && xhr.status < 300) {
+                                // dispatch custom event once cookie loading request has returned
+                                const event = new CustomEvent("updateCookieConsent", {
+                                    detail: {
+                                        groups: updatedCookieGroups.split(',')
+                                    }
+                                });
+                                document.dispatchEvent(event);
+                            }
+                        };
+                        xhr.send();
+                        // set ccokies on additional hosts using helper image
                         if (typeof (additionalHostLinks) != 'undefined' && additionalHostLinks != null) {
                             additionalHostLinks.split(',').forEach(function (url) {
                                 let img = document.createElement("img");
@@ -50,12 +61,7 @@ import Cookies from "js-cookie";
                                 document.body.appendChild(img);
                             });
                         }
-                        const event = new CustomEvent("updateCookieConsent", {
-                            detail: {
-                                groups: updatedCookieGroups.split(',')
-                            }
-                        });
-                        document.dispatchEvent(event);
+                        // hide popup
                         if (typeof (popup) != 'undefined' && popup != null) {
                             popup.style.display = 'none';
                         }